fix(FR-3402): render published announcements through the editor's markdown pipeline
The announcement editor's Preview pane and the published announcement alert
used two different markdown renderers. The preview ran `react-markdown` with
`remark-gfm` / `remark-math` / `rehype-katex` and a stylesheet that gave
blockquotes an accent bar and background; the alert ran `markdown-to-jsx` with
no stylesheet at all. A blockquote therefore looked correct while editing and
came out as plain text once published, and the same divergence applied to
tables, task lists, and math.

Extract a single `AnnouncementMarkdown` component holding the renderer, the
plugin set, and the typography styles, and render both views through it, so
the preview's "this is what will be published" contract holds by construction.
