+++ react/src/components/AnnouncementAlert.tsx
@@ -5,12 +5,12 @@
 import { useCurrentUserRole } from '../hooks/backendai';
 import { useSuspenseGetAnnouncement } from '../hooks/useSuspenseGetAnnouncement';
 import AnnouncementEditModal from './AnnouncementEditModal';
+import AnnouncementMarkdown from './AnnouncementMarkdown';
 import { useToggle } from 'ahooks';
-import { Button, theme } from 'antd';
+import { Button } from 'antd';
 import { BAIAlert, BAIAlertProps, BAIUnmountAfterClose } from 'backend.ai-ui';
 import * as _ from 'lodash-es';
 import { SquarePenIcon } from 'lucide-react';
-import Markdown from 'markdown-to-jsx';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -19,7 +19,6 @@ const AnnouncementAlert: React.FC<Props> = ({ ...otherProps }) => {
   'use memo';
 
   const { t } = useTranslation();
-  const { token } = theme.useToken();
   const userRole = useCurrentUserRole();
   const isSuperAdmin = userRole === 'superadmin';
   const [isEditOpen, { toggle: toggleEditModal }] = useToggle(false);
@@ -29,22 +28,7 @@ const AnnouncementAlert: React.FC<Props> = ({ ...otherProps }) => {
     <>
       <BAIAlert
         description={
-          <div style={{ marginBottom: token.marginSM * -1 }}>
-            <Markdown
-              options={{
-                overrides: {
-                  p: {
-                    props: {
-                      style: { marginTop: 0, marginBottom: token.marginSM },
-                    },
-                  },
-                },
-              }}
-            >
-              {/* trailing <p> collapses the last paragraph's bottom margin */}
-              {announcement.message + '<p></p>'}
-            </Markdown>
-          </div>
+          <AnnouncementMarkdown>{announcement.message}</AnnouncementMarkdown>
         }
         action={
           isSuperAdmin ? (
