+++ react/src/components/AnnouncementEditModal.tsx
@@ -5,8 +5,8 @@
 import { useSuspendedBackendaiClient } from '../hooks';
 import { useTanMutation, useTanQuery } from '../hooks/reactQueryAlias';
 import { announcementQueryOptions } from '../hooks/useSuspenseGetAnnouncement';
+import AnnouncementMarkdown from './AnnouncementMarkdown';
 import BAICodeEditor from './BAICodeEditor';
-import { SyntaxHighlighter } from './Chat/SyntaxHighlighter';
 import {
   BoldOutlined,
   CodeOutlined,
@@ -37,153 +37,13 @@ import {
   useErrorMessageResolver,
   useBAILogger,
 } from 'backend.ai-ui';
-// `rehype-katex` does not import the CSS file, so we need to import it manually.
-import 'katex/dist/katex.min.css';
 import { useCallback, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import Markdown from 'react-markdown';
-import rehypeKatex from 'rehype-katex';
-import remarkGfm from 'remark-gfm';
-import remarkMath from 'remark-math';
 
 type MonacoEditorInstance = Parameters<OnMount>[0];
 type MonacoNamespace = Parameters<OnMount>[1];
 
 const useStyles = createStyles(({ css, token }) => ({
-  // velog-style reading typography for the live preview: comfortable line
-  // height, bordered headings, accented blockquotes, GitHub-flavored tables.
-  markdownPreview: css`
-    color: ${token.colorText};
-    font-size: ${token.fontSize}px;
-    line-height: 1.7;
-    word-break: break-word;
-
-    & > *:first-child {
-      margin-top: 0;
-    }
-    & > *:last-child {
-      margin-bottom: 0;
-    }
-
-    h1,
-    h2,
-    h3,
-    h4,
-    h5,
-    h6 {
-      margin: 1.5em 0 0.75em;
-      font-weight: ${token.fontWeightStrong};
-      line-height: 1.4;
-    }
-    h1 {
-      font-size: 1.9em;
-      padding-bottom: 0.3em;
-      border-bottom: 1px solid ${token.colorBorderSecondary};
-    }
-    h2 {
-      font-size: 1.5em;
-      padding-bottom: 0.3em;
-      border-bottom: 1px solid ${token.colorBorderSecondary};
-    }
-    h3 {
-      font-size: 1.25em;
-    }
-    h4 {
-      font-size: 1em;
-    }
-
-    p {
-      margin: 0 0 1em;
-    }
-
-    a {
-      color: ${token.colorLink};
-      text-decoration: none;
-    }
-    a:hover {
-      text-decoration: underline;
-    }
-
-    img {
-      max-width: 100%;
-      border-radius: ${token.borderRadius}px;
-    }
-
-    /* Restore list markers: the global reset in resources/webui.css sets
-       \`ul { list-style-type: none }\`, which otherwise hides bullets here. */
-    ul,
-    ol {
-      margin: 0 0 1em;
-      padding-left: 1.6em;
-    }
-    ul {
-      list-style: disc;
-    }
-    ul ul {
-      list-style: circle;
-    }
-    ol {
-      list-style: decimal;
-    }
-    li {
-      margin: 0.25em 0;
-    }
-    li > p {
-      margin: 0;
-    }
-    li > input[type='checkbox'] {
-      margin-right: 0.4em;
-    }
-
-    blockquote {
-      margin: 0 0 1em;
-      padding: 0.4em 1em;
-      color: ${token.colorTextSecondary};
-      border-left: 4px solid ${token.colorPrimary};
-      background: ${token.colorFillQuaternary};
-      border-radius: ${token.borderRadiusSM}px;
-    }
-    blockquote > *:last-child {
-      margin-bottom: 0;
-    }
-
-    hr {
-      margin: 1.5em 0;
-      border: none;
-      border-top: 1px solid ${token.colorBorderSecondary};
-    }
-
-    /* Inline code only; fenced blocks are rendered by SyntaxHighlighter. */
-    :not(pre) > code {
-      padding: 0.15em 0.4em;
-      font-family: ${token.fontFamilyCode};
-      font-size: 0.9em;
-      background: ${token.colorFillSecondary};
-      border-radius: ${token.borderRadiusSM}px;
-    }
-    pre {
-      margin: 0 0 1em;
-      border-radius: ${token.borderRadius}px;
-      overflow: auto;
-    }
-
-    table {
-      width: 100%;
-      margin: 0 0 1em;
-      border-collapse: collapse;
-      font-size: 0.95em;
-    }
-    th,
-    td {
-      padding: ${token.paddingXS}px ${token.paddingSM}px;
-      border: 1px solid ${token.colorBorderSecondary};
-    }
-    th {
-      background: ${token.colorFillTertiary};
-      font-weight: ${token.fontWeightStrong};
-      text-align: left;
-    }
-  `,
   toolbar: css`
     padding: ${token.paddingXXS}px ${token.paddingXS}px;
     border: 1px solid ${token.colorBorder};
@@ -220,7 +80,6 @@ const AnnouncementEditModal: React.FC<AnnouncementEditModalProps> = ({
 
   const { t } = useTranslation();
   const { token } = theme.useToken();
-  const { styles } = useStyles();
   const { message: appMessage, modal } = App.useApp();
   const { logger } = useBAILogger();
   const { getErrorMessage } = useErrorMessageResolver();
@@ -400,8 +259,7 @@ const AnnouncementEditModal: React.FC<AnnouncementEditModalProps> = ({
             <Typography.Text strong>
               {t('summary.AnnouncementPreview')}
             </Typography.Text>
-            <div
-              className={styles.markdownPreview}
+            <AnnouncementMarkdown
               style={{
                 border: `1px solid ${token.colorBorder}`,
                 borderRadius: token.borderRadius,
@@ -413,36 +271,8 @@ const AnnouncementEditModal: React.FC<AnnouncementEditModalProps> = ({
                 overflow: 'auto',
               }}
             >
-              <Markdown
-                remarkPlugins={[remarkGfm, remarkMath]}
-                rehypePlugins={[rehypeKatex]}
-                components={{
-                  // Fenced code blocks: render through the shared shiki-based
-                  // highlighter (theme-aware). Inline code keeps the default
-                  // <code> and is styled via CSS.
-                  pre({ children }) {
-                    const codeElement = Array.isArray(children)
-                      ? children[0]
-                      : children;
-                    const className: string =
-                      // @ts-ignore - react-markdown passes the <code> element here
-                      codeElement?.props?.className ?? '';
-                    const match = /language-(\w+)/.exec(className);
-                    const content = String(
-                      // @ts-ignore
-                      codeElement?.props?.children ?? '',
-                    ).replace(/\n$/, '');
-                    return (
-                      <SyntaxHighlighter language={match?.[1] ?? 'txt'}>
-                        {content}
-                      </SyntaxHighlighter>
-                    );
-                  },
-                }}
-              >
-                {message}
-              </Markdown>
-            </div>
+              {message}
+            </AnnouncementMarkdown>
           </BAIFlex>
         </BAIFlex>
       )}
